bloke_ai: match ControlPeople, build-object save/load; no casts or struct keywords left
- ControlPeople: plain list walk; low-level AI called from both the on-ride and normal
  branches (target duplicates the field_e check)
- DAT_006664f8 is BuildObj[256] (ride pointer, saved as its LLIDB index); save/load use
  int-index loops (target compares signed); array and DAT_006670f8 initialised so they
  stay adjacent for the one-past-the-end compare
- GeteListPtr returns Element *, FindeIneList takes Element **; DAT_00669200 is Element **
- FirstBloke is Bloke *; DoLowLevelAI takes Bloke *; FUN_004700c0 takes void *
  (DAT_007fdf8c, the popup's object); Bloke.field_78 is short

+++ src/legoland/popupinfo.c
@@ -579,7 +579,7 @@ LEGO_EXPORT void PopUpInfoSetUp(int param_1, void *param_2, unsigned int param_3
             }
         } else if (param_1 == 0x306) {
             DAT_007fdf9c = param_1;
-            DAT_007fdf8c = (unsigned int)param_2;
+            DAT_007fdf8c = param_2;
             DAT_007fdf90 = *(unsigned int *)(*(int *)((char *)param_2 + 4) + 0x1c);
             DAT_007fdf94 = *(unsigned int *)(*(int *)((char *)param_2 + 4) + 0x20);
             return;
